+++ packages/dnr-rulesets/src/lib/assets/loader.ts
@@ -7,13 +7,18 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 /**
  * Api for loading assets.
+ *
+ * TODO: Rename to `AssetsManager`, because it is not load anything.
  */
 export class AssetsLoader {
     /**
      * Download filters to {@link dest} path.
      *
      * @param dest Path to download filters.
+     *
      * @returns Promise that resolves when filters are downloaded.
+     *
+     * TODO: Make this method static.
      */
     public async load(dest: string): Promise<void> {
         const to = path.resolve(process.cwd(), dest);