+++ packages/dnr-rulesets/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1-alpha.0] - 2025-05-28
+
+### Added
+- New `watch` command to watch for changes in the filters
+  and automatically rebuild DNR.
+
+[3.0.1-alpha.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/dnr-rulesets-v3.0.1-alpha.0
+
 ## [3.0.0-alpha.7] - 2025-05-19
 
 ### Changed